Make wok_smtp_mailer:send always return an ok/error tuple
When a callback is passed (always the case when using wok_mailer),
gen_smtp returns a pid instead of a {ok, ...} tuple. This makes it
difficult to match for success.

We now wrap such results in an {ok, Pid} tuple. gen_smtp start errors
are also returned in addition to being passed to the callback.